requirements.txt
# package dependencies
click
distro
GitPython
pyyaml

# dev dependencies
parameterized
pytest
